backend/requirements.txt
amqp==5.2.0
asgiref==3.8.1
attrs==23.2.0
billiard==4.2.1
black==24.4.2
boto3==1.34.110
botocore==1.34.110
cachetools==5.5.0
cattrs==23.2.3
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
cryptography==43.0.1
decorator==5.1.1
distlib==0.3.8
Django==5.0.6
django-celery-beat==2.7.0
django-cors-headers==4.5.0
django-storages==1.14.3
django-timezone-field==7.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
filelock==3.14.0
grpcio==1.66.2
idna==3.10
jmespath==1.0.1
jsonpath-rw==1.4.0
kombu==5.4.2
lsprotocol==2023.0.1
mmh3==5.0.1
mypy-extensions==1.0.0
orjson==3.10.7
packaging==23.1
pathspec==0.12.1
pbr==6.0.0
platformdirs==4.2.2
ply==3.11
prompt-toolkit==3.0.36
protobuf==3.20.0
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyfiglet==1.0.2
pygls==1.3.1
PyJWT==2.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
questionary==2.0.1
regex==2024.9.11
requests==2.32.3
ruff==0.4.6
ruff-lsp==0.0.53
s3transfer==0.10.1
service-identity==24.1.0
setuptools==75.1.0
simplejson==3.19.3
six==1.16.0
sqlparse==0.5.0
stevedore==5.2.0
tb-mqtt-client==1.10.5
tb-paho-mqtt-client==1.6.3
termcolor==2.4.0
typing_extensions==4.12.0
tzdata==2024.2
urllib3==1.26.15
vine==5.1.0
virtualenv==20.26.2
virtualenv-clone==0.5.7
virtualenvwrapper==6.1.0
wcwidth==0.2.13
